# Optimize Write Signature calculation